Use single precision instead

+++ c/hexagon/inc/ecg_kalman.h
@@ -14,8 +14,8 @@ extern "C" {
 #define N 5 // Averaging for state vector
 #define ECG_LEADS 12
 
-int preprocess_ecg_data(double **data, int rows, double **preprocessed_output, int *ecg_complex_length);
-int process_kalman(double **data, int rows, int cols, int ecg_complex_length, double *output);
+int preprocess_ecg_data(float **data, int rows, float **preprocessed_output, int *ecg_complex_length);
+int process_kalman(float **data, int rows, int cols, int ecg_complex_length, float *output);
 
 #ifdef __cplusplus
 }
